fix(utils): extract censor_tokens to a common function
