+++ aperturedb/CommonLibrary.py
@@ -300,20 +300,7 @@ def execute_query(client: Connector, query: Commands,
     logger.debug(f"Query={query}")
     r, b = client.query(query, blobs)
 
-    # Censor token information in the response for logging
-    def censor_tokens(response):
-        import copy
-        if isinstance(response, list):
-            censored = copy.deepcopy(response)
-            for item in censored:
-                if isinstance(item, dict) and "Authenticate" in item:
-                    auth = item["Authenticate"]
-                    for k in ["refresh_token", "session_token"]:
-                        if k in auth and isinstance(auth[k], str) and len(auth[k]) > 8:
-                            auth[k] = auth[k][:4] + "..." + auth[k][-4:]
-            return censored
-        return response
-
+    from aperturedb.Utils import censor_tokens
     censored_r = censor_tokens(r)
     logger.debug(f"Response={censored_r}")
